+++ src/main/java/bumblebee/xchangepass/domain/refresh/dto/RefreshTokenRequest.java
@@ -1,4 +1,4 @@
-package bumblebee.xchangepass.global.security.refresh.dto;
+package bumblebee.xchangepass.domain.refresh.dto;
 
 public record RefreshTokenRequest(String refreshToken) {
